Extract ISO_8859_7 to its own file
