+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/Charset.kt
@@ -58,141 +58,6 @@ public sealed class Charset(
   //  /** Page 14 [CP737: Greek] */
   //  CP737(14, TODO()),
 
-  /** Page 15 [ISO8859-7: Greek] */
-  public data object ISO_8859_7 : Charset(15, "ISO-8859-7") {
-    override val mapping: Map<Char, Int> =
-        mapOf(
-            '\u0080' to 0x80,
-            '\u0081' to 0x81,
-            '\u0082' to 0x82,
-            '\u0083' to 0x83,
-            '\u0084' to 0x84,
-            '\u0085' to 0x85,
-            '\u0086' to 0x86,
-            '\u0087' to 0x87,
-            '\u0088' to 0x88,
-            '\u0089' to 0x89,
-            '\u008A' to 0x8A,
-            '\u008B' to 0x8B,
-            '\u008C' to 0x8C,
-            '\u008D' to 0x8D,
-            '\u008E' to 0x8E,
-            '\u008F' to 0x8F,
-            '\u0090' to 0x90,
-            '\u0091' to 0x91,
-            '\u0092' to 0x92,
-            '\u0093' to 0x93,
-            '\u0094' to 0x94,
-            '\u0095' to 0x95,
-            '\u0096' to 0x96,
-            '\u0097' to 0x97,
-            '\u0098' to 0x98,
-            '\u0099' to 0x99,
-            '\u009A' to 0x9A,
-            '\u009B' to 0x9B,
-            '\u009C' to 0x9C,
-            '\u009D' to 0x9D,
-            '\u009E' to 0x9E,
-            '\u009F' to 0x9F,
-            '\u00A0' to 0xA0,
-            '\u2018' to 0xA1,
-            '\u2019' to 0xA2,
-            '\u00A3' to 0xA3,
-            '\u20AC' to 0xA4,
-            '\u20AF' to 0xA5,
-            '\u00A6' to 0xA6,
-            '\u00A7' to 0xA7,
-            '\u00A8' to 0xA8,
-            '\u00A9' to 0xA9,
-            '\u037A' to 0xAA,
-            '\u00AB' to 0xAB,
-            '\u00AC' to 0xAC,
-            '\u00AD' to 0xAD,
-            '\uFFFD' to 0xAE,
-            '\u2015' to 0xAF,
-            '\u00B0' to 0xB0,
-            '\u00B1' to 0xB1,
-            '\u00B2' to 0xB2,
-            '\u00B3' to 0xB3,
-            '\u0384' to 0xB4,
-            '\u0385' to 0xB5,
-            '\u0386' to 0xB6,
-            '\u00B7' to 0xB7,
-            '\u0388' to 0xB8,
-            '\u0389' to 0xB9,
-            '\u038A' to 0xBA,
-            '\u00BB' to 0xBB,
-            '\u038C' to 0xBC,
-            '\u00BD' to 0xBD,
-            '\u038E' to 0xBE,
-            '\u038F' to 0xBF,
-            '\u0390' to 0xC0,
-            '\u0391' to 0xC1,
-            '\u0392' to 0xC2,
-            '\u0393' to 0xC3,
-            '\u0394' to 0xC4,
-            '\u0395' to 0xC5,
-            '\u0396' to 0xC6,
-            '\u0397' to 0xC7,
-            '\u0398' to 0xC8,
-            '\u0399' to 0xC9,
-            '\u039A' to 0xCA,
-            '\u039B' to 0xCB,
-            '\u039C' to 0xCC,
-            '\u039D' to 0xCD,
-            '\u039E' to 0xCE,
-            '\u039F' to 0xCF,
-            '\u03A0' to 0xD0,
-            '\u03A1' to 0xD1,
-            '\uFFFD' to 0xD2,
-            '\u03A3' to 0xD3,
-            '\u03A4' to 0xD4,
-            '\u03A5' to 0xD5,
-            '\u03A6' to 0xD6,
-            '\u03A7' to 0xD7,
-            '\u03A8' to 0xD8,
-            '\u03A9' to 0xD9,
-            '\u03AA' to 0xDA,
-            '\u03AB' to 0xDB,
-            '\u03AC' to 0xDC,
-            '\u03AD' to 0xDD,
-            '\u03AE' to 0xDE,
-            '\u03AF' to 0xDF,
-            '\u03B0' to 0xE0,
-            '\u03B1' to 0xE1,
-            '\u03B2' to 0xE2,
-            '\u03B3' to 0xE3,
-            '\u03B4' to 0xE4,
-            '\u03B5' to 0xE5,
-            '\u03B6' to 0xE6,
-            '\u03B7' to 0xE7,
-            '\u03B8' to 0xE8,
-            '\u03B9' to 0xE9,
-            '\u03BA' to 0xEA,
-            '\u03BB' to 0xEB,
-            '\u03BC' to 0xEC,
-            '\u03BD' to 0xED,
-            '\u03BE' to 0xEE,
-            '\u03BF' to 0xEF,
-            '\u03C0' to 0xF0,
-            '\u03C1' to 0xF1,
-            '\u03C2' to 0xF2,
-            '\u03C3' to 0xF3,
-            '\u03C4' to 0xF4,
-            '\u03C5' to 0xF5,
-            '\u03C6' to 0xF6,
-            '\u03C7' to 0xF7,
-            '\u03C8' to 0xF8,
-            '\u03C9' to 0xF9,
-            '\u03CA' to 0xFA,
-            '\u03CB' to 0xFB,
-            '\u03CC' to 0xFC,
-            '\u03CD' to 0xFD,
-            '\u03CE' to 0xFE,
-            '\uFFFD' to 0xFF,
-        )
-  }
-
   /** Page 16 [Windows-1252] */
   public data object Windows1252 : Charset(16, "windows-1252") {
     override val mapping: Map<Char, Int> =
